requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pymorphy3>=1.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
numpy>=1.21.0
striprtf>=0.0.15
python-multipart>=0.0.5
pydantic>=1.8.0